+++ src/main/java/cart/WebMvcConfig.java
@@ -1,8 +1,8 @@
 package cart;
 
 import cart.dao.MemberDao;
-import cart.ui.MemberArgumentResolver;
-import cart.ui.PageArgumentResolver;
+import cart.ui.argument_resolver.MemberArgumentResolver;
+import cart.ui.argument_resolver.PageArgumentResolver;
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
