+++ src/main/java/cart/ui/controller/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package cart.ui;
+package cart.ui.controller;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
