+++ src/main/java/cart/ui/controller/PointController.java
@@ -1,4 +1,4 @@
-package cart.ui;
+package cart.ui.controller;
 
 import cart.domain.Member;
 import cart.domain.Point;
