+++ src/main/java/cart/ui/controller/ProductController.java
@@ -1,4 +1,4 @@
-package cart.ui;
+package cart.ui.controller;
 
 import cart.application.ProductService;
 import cart.dto.ProductRequest;
