+++ src/test/java/cart/ui/OrderControllerTest.java
@@ -19,7 +19,8 @@ import cart.dto.OrderDetailResponse;
 import cart.dto.OrderItemResponse;
 import cart.dto.ProductResponse;
 import cart.entity.MemberEntity;
-import cart.ui.pageable.Page;
+import cart.ui.controller.OrderController;
+import cart.ui.paging.Page;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.time.LocalDateTime;
