+++ src/test/java/cart/ui/PointControllerTest.java
@@ -8,6 +8,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import cart.dao.MemberDao;
 import cart.entity.MemberEntity;
+import cart.ui.controller.PointController;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
